requirements.txt
PyQt5>=5.15.0
requests>=2.25.0
python-whois>=0.8.0
dnspython>=2.1.0